+++ app/api/admin/grupos/route.ts
@@ -154,7 +154,7 @@ export async function GET(req: NextRequest) {
                   transacoes: transacoesRes.rows.map((t) => ({
                     id: t.id,
                     data: t.data,
-                    quantidade: parseInt(t.quantidade) || 0,
+                    quantidade: parseFloat(t.quantidade) || 0,
                     valorUnitario: parseFloat(t.valor_unitario) || 0,
                   })),
                 }
